Rewrite SQL and DBML rendering
