+++ pydbml/__init__.py
@@ -1,6 +1,3 @@
-from . import classes
+from . import _classes
 from .parser import PyDBML
 from .database import Database
-from pydbml.constants import MANY_TO_ONE
-from pydbml.constants import ONE_TO_MANY
-from pydbml.constants import ONE_TO_ONE
