+++ pydbml/classes/__init__.py
@@ -1,10 +1,11 @@
-from .column import Column
-from .enum import Enum
-from .enum import EnumItem
-from .expression import Expression
-from .index import Index
-from .note import Note
-from .project import Project
-from .reference import Reference
-from .table import Table
-from .table_group import TableGroup
+from .._classes.column import Column
+from .._classes.enum import Enum
+from .._classes.enum import EnumItem
+from .._classes.expression import Expression
+from .._classes.index import Index
+from .._classes.note import Note
+from .._classes.project import Project
+from .._classes.reference import Reference
+from .._classes.sticky_note import StickyNote
+from .._classes.table import Table
+from .._classes.table_group import TableGroup
