+++ pydbml/tools.py
@@ -2,26 +2,11 @@ import re
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:  # pragma: no cover
-    from .classes import Note
+    pass
 
 
 def comment(val: str, comb: str) -> str:
     return '\n'.join(f'{comb} {cl}' for cl in val.split('\n')) + '\n'
-
-
-def comment_to_dbml(val: str) -> str:
-    return comment(val, '//')
-
-
-def comment_to_sql(val: str) -> str:
-    return comment(val, '--')
-
-
-def note_option_to_dbml(note: 'Note') -> str:
-    if '\n' in note.text:
-        return f"note: '''{note._prepare_text_for_dbml()}'''"
-    else:
-        return f"note: '{note._prepare_text_for_dbml()}'"
 
 
 def indent(val: str, spaces=4) -> str:
